solution/src/main/java/com/alanjwilliams/engage/util/JsonLocalDateSerializer.java
package com.alanjwilliams.engage.util;

import com.fasterxml.jackson.core.JsonGenerator;
import com.fasterxml.jackson.core.JsonProcessingException;
import com.fasterxml.jackson.databind.JsonSerializer;
import com.fasterxml.jackson.databind.SerializerProvider;

import java.io.IOException;
import java.time.LocalDate;
import java.time.format.DateTimeFormatter;

/**
 * JSON LocalDate Serializer.
 */
public class JsonLocalDateSerializer extends JsonSerializer<LocalDate> {

    /**
     * DateTimeFormatter.
     */
    private static final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("dd/MM/uuuu");

    @Override
    public void serialize(LocalDate localDate, JsonGenerator jsonGenerator, SerializerProvider serializerProvider) throws IOException, JsonProcessingException {
        jsonGenerator.writeString(localDate.format(dateTimeFormatter));
    }
}
